- enhancement: [Email module - backend] EmailAccount- and related information is now deleted entirely if the belonging folder is of the type root; "accounts_root" flagged information will be deleted if no more email items belonging to the accounts folders can be found

+++ src/corelib/php/library/Conjoon/Modules/Groupware/Email/Account/Model/Account.php
@@ -93,8 +93,11 @@ class Conjoon_Modules_Groupware_Email_Account_Model_Account
     }
 
     /**
-     * Completely removes the account and all associated data - that means
-     * that all information in folders_accounts will be removed, too.
+     * Sets the is_deleted flag of the account to "0", which basically means that
+     * this account is not active anymore. However, in order for older email items
+     * to still work properly, folder mappings have to remain, until no
+     * item for this account is found in the database. After that, the account may
+     * be removed entirely.
      *
      * @param integer $accountId
      * @param integer $userId
@@ -111,20 +114,71 @@ class Conjoon_Modules_Groupware_Email_Account_Model_Account
         }
 
         /**
+         * @see Conjoon_Modules_Groupware_Email_Folder_Model_Folder
+         */
+        require_once 'Conjoon/Modules/Groupware/Email/Folder/Model/Folder.php';
+
+        $folderModel = new Conjoon_Modules_Groupware_Email_Folder_Model_Folder();
+
+        $rootId = $folderModel->getRootFolderId($accountId, $userId);
+
+        /**
          * @see Conjoon_Modules_Groupware_Email_Folder_Model_FoldersAccounts
          */
         require_once 'Conjoon/Modules/Groupware/Email/Folder/Model/FoldersAccounts.php';
 
         $foldersAccounts = new Conjoon_Modules_Groupware_Email_Folder_Model_FoldersAccounts();
 
-        // delete the account/folder map here
-        $foldersAccounts->deleteForAccountId($accountId);
+        $folders = $foldersAccounts->getFolderIdsForAccountId($accountId);
 
-        $where   = $this->getAdapter()->quoteInto('id = ?', $accountId, 'INTEGER');
-        $deleted = $this->delete($where);
+        // rootID found - delete all!
+        if ($rootId) {
+            // no folders or root - we can remove the account entirely
+            $where   = $this->getAdapter()->quoteInto('id = ?', $accountId, 'INTEGER');
+            $deleted = $this->delete($where);
 
+            for ($i = 0, $len = count($folders); $i < $len) {
+                $folderModel->deleteFolder($folders[$i], $userId);
+            }
 
-        return $deleted;
+            return $deleted;
+        } else {
+            // no root id. Check if there are any items still in folders
+            // belonging to the account. If that is the case, DO NOT remove the
+            // account from the data storage
+            /**
+             * @see Conjoon_Modules_Groupware_Email_Item_Model_Item
+             */
+            require_once 'Conjoon/Modules/Groupware/Email/Item/Model/Item.php';
+
+            $itemModel = new Conjoon_Modules_Groupware_Email_Item_Model_Item();
+            $del = true;
+            // check if for any found folder id still exists an email item
+            for ($i = 0, $len = count($folders); $i < $len; $i++) {
+                $folderId = $folders[$i];
+                $count = $itemModel->getEmailItemCountForFolder($folders[$i]);
+                if ($count > 0) {
+                    $del = false;
+                    break;
+                }
+            }
+
+            if ($del) {
+                // remove account-data entirely - as for accounts_root data specified
+                $where   = $this->getAdapter()->quoteInto('id = ?', $accountId, 'INTEGER');
+                $deleted = $this->delete($where);
+
+                // delete account-mappings
+                $foldersAccounts->deleteForAccountId($accountId);
+
+                return $deleted;
+            } else {
+                // update account to is_deleted = 1
+                $where = $this->getAdapter()->quoteInto('id = ?', $accountId, 'INTEGER');
+                return $this->update(array('is_deleted' => 1), $where);
+            }
+        }
+
     }
 
     /**
